refactor: move curric download selection into its own component
