+++ src/components/CurriculumComponents/CurriculumDownloadView/SignedInFlow.tsx
@@ -1,23 +1,11 @@
-import {
-  OakBox,
-  OakFlex,
-  OakPrimaryButton,
-  OakDownloadCard,
-  OakHeading,
-  OakTagFunctional,
-} from "@oaknational/oak-components";
+import { OakBox, OakFlex, OakPrimaryButton } from "@oaknational/oak-components";
 import { useState } from "react";
 import { useUser } from "@clerk/nextjs";
-import { uniq } from "lodash";
 
 import Terms from "../OakComponentsKitchen/Terms";
 
-import {
-  DOWNLOAD_TYPES,
-  DownloadType,
-  School,
-  assertValidDownloadType,
-} from "./helper";
+import { DOWNLOAD_TYPES, DownloadType, School } from "./helper";
+import { CurriculumDownloadSelection } from "./CurriculumDownloadSelection";
 
 import { CurriculumDownloadViewProps } from ".";
 
@@ -82,66 +70,10 @@ export default function SignedInFlow({ onSubmit, schools }: SignedInFlowProps) {
       $alignItems={"flex-start"}
     >
       <Box $width={["100%", 510]} $textAlign={"left"}>
-        <OakFlex $flexDirection={"column"} $gap={"space-between-s"}>
-          <OakHeading
-            tag="h3"
-            $font={["heading-5"]}
-            data-testid="download-heading"
-          >
-            Curriculum resources
-          </OakHeading>
-          <OakFlex $flexDirection={"column"} $gap={"space-between-s"}>
-            {DOWNLOAD_TYPES.map((download) => {
-              const isChecked = downloadTypes.includes(download.id);
-              const isEditable = download.subTitle?.includes("accessible");
-
-              return (
-                <OakDownloadCard
-                  key={download.id}
-                  id={download.id}
-                  data-testid="resourceCard"
-                  value={download.id}
-                  name="curriculum-download"
-                  titleSlot={download.label}
-                  checked={isChecked}
-                  onChange={(e) => {
-                    const downloadType = assertValidDownloadType(
-                      e.target.value,
-                    );
-                    let newDownloadTypes: DownloadType[];
-                    if (e.target.checked) {
-                      newDownloadTypes = uniq([...downloadTypes, downloadType]);
-                    } else {
-                      newDownloadTypes = downloadTypes.filter(
-                        (id) => id !== downloadType,
-                      );
-                    }
-                    setDownloadTypes(newDownloadTypes);
-                  }}
-                  formatSlot={
-                    <>
-                      {download.subTitle}
-                      {isEditable && (
-                        <OakTagFunctional
-                          key="tag"
-                          $ml={"space-between-ssx"}
-                          $display="inline"
-                          $color={"text-primary"}
-                          $font={"heading-light-7"}
-                          $ph={"inner-padding-ssx"}
-                          $pv={"inner-padding-ssx"}
-                          label="Editable"
-                          $background={"bg-decorative2-main"}
-                        />
-                      )}
-                    </>
-                  }
-                  iconName={download.icon}
-                />
-              );
-            })}
-          </OakFlex>
-        </OakFlex>
+        <CurriculumDownloadSelection
+          downloadTypes={downloadTypes}
+          onChange={setDownloadTypes}
+        />
         <OakBox $mt="space-between-m">
           <Terms />
         </OakBox>
